Renamed Password/SMS folders to improve naming consistency
