+++ packages/apollos-ui-auth/src/Password/Signup/index.js
